# public/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brainswarming API",
        "description": "API for the Brainswarming platform - a SaaS platform for teams to anonymously or openly submit, evaluate, and prioritize improvement ideas.",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/teams/{teamId}/entries": {
            "get": {
                "tags": [
                    "Entries"
                ],
                "summary": "Get all entries for a team",
                "description": "Returns a list of all entries for the specified team, ordered by priority",
                "operationId": "getTeamEntries",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "area",
                        "in": "query",
                        "description": "Filter entries by area",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "entries": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Entry"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Team not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Entries"
                ],
                "summary": "Create a new entry for a team",
                "description": "Creates a new entry for the specified team",
                "operationId": "storeTeamEntry",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "problem",
                                    "solution",
                                    "area",
                                    "effort",
                                    "monetary_explanation"
                                ],
                                "properties": {
                                    "problem": {
                                        "type": "string",
                                        "example": "Slow customer response time"
                                    },
                                    "solution": {
                                        "type": "string",
                                        "example": "Implement automated response system"
                                    },
                                    "area": {
                                        "type": "string",
                                        "example": "Customer Service"
                                    },
                                    "time_saved_per_year": {
                                        "type": "integer",
                                        "example": 500,
                                        "nullable": true
                                    },
                                    "gross_profit_per_year": {
                                        "type": "integer",
                                        "example": 10000,
                                        "nullable": true
                                    },
                                    "effort": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ],
                                        "example": "medium"
                                    },
                                    "monetary_explanation": {
                                        "type": "string",
                                        "example": "Saves 500 hours per year at $20/hour"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://example.com/docs",
                                        "nullable": true
                                    },
                                    "anonymous": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "manual_override_prio": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "entry": {
                                            "$ref": "#/components/schemas/Entry"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Team not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/teams/{teamId}/entries/deleted": {
            "get": {
                "tags": [
                    "Entries"
                ],
                "summary": "Get all deleted entries for a team",
                "description": "Returns a list of all soft-deleted entries for the specified team",
                "operationId": "getDeletedTeamEntries",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "entries": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Entry"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Team not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/teams/{teamId}/entries/export": {
            "get": {
                "tags": [
                    "Entries"
                ],
                "summary": "Export team entries",
                "description": "Exports all entries for the specified team in CSV format",
                "operationId": "exportTeamEntries",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "ID of the team",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file containing all entries",
                        "headers": {
                            "Content-Type": {
                                "description": "text/csv",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Disposition": {
                                "description": "attachment; filename=entries.csv",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Team not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}